implemented service

+++ src/main/java/com/cg/dto/BusBookingDto.java
@@ -5,6 +5,6 @@ import java.util.List;
 
 import com.cg.entity.Passenger;
 
-public record BusBookingDto(Long scheduleId, Integer custId, LocalDate bookingDt, List<Passenger> passengers) {
+public record BusBookingDto(Long scheduleId, Long custId, LocalDate bookingDt, List<Passenger> passengers) {
 
 }
